requirements.txt
uvicorn
pydantic-settings
jieba3k>=0.35.1
wordcloud>=1.9.4
scikit-learn>=1.7.2
sqlalchemy>=2.0.44
fastapi>=0.122.0
matplotlib>=3.10.7
pandas>=2.3.3
numpy>=2.3.5
pydantic_settings>=2.12.0
asyncpg>=0.31.0
